+++ nose2/tests/functional/support/scenario/junitxml/skip_reason/test_junitxml_skip_reason.py
@@ -1,0 +1,8 @@
+import unittest
+
+
+class Test(unittest.TestCase):
+
+    @unittest.skip('ohai')
+    def test(self):
+        pass